+++ examples/connect/main.cjs
@@ -1,0 +1,8 @@
+const { Client } = require('ts-postgres');
+const client = new Client();
+module.exports = client.connect().then(
+  (info) => {
+    console.log("Encrypted: " + info.encrypted);
+    return client.end();
+  }
+);