+++ src/protocol.ts
@@ -1,16 +1,18 @@
-import { Socket } from 'net';
-import { Writable } from 'stream';
-import { ElasticBuffer } from './buffer';
-import { postgresqlErrorCodes } from './errors';
-import { sign } from './sasl';
-import { sum } from './utils';
+import { Buffer } from 'node:buffer';
+import { Socket } from 'node:net';
+import { Writable } from 'node:stream';
+import { ElasticBuffer } from './buffer.js';
+import { postgresqlErrorCodes } from './errors.js';
+import { sign } from './sasl.js';
+import { sum } from './utils.js';
 import {
     arrayDataTypeMapping,
     isPoint,
+    BufferEncoding,
     DataFormat,
     DataType,
     ValueTypeReader
-} from './types';
+} from './types.js';
 
 const arrayMask = 1 << 31;
 const readerMask = 1 << 29;
@@ -278,7 +280,6 @@ export function readRowDescription(
         names: names
     }
 }
-
 
 export function readRowData(
     buffer: Buffer,
@@ -1092,7 +1093,7 @@ export class Writer {
     }
 
     startup(settings: StartupConfiguration) {
-        const data = [];
+        const data: string[] = [];
         const options = {
             "user": settings.user,
             "database": settings.database,