+++ CHANGES.md
@@ -1,5 +1,11 @@
 In the next release ...
 
+- Add support for ESM modules.
+
+- The database host, port and connection timeout options can now be
+  specified directly for `connect` (taking priority over the
+  provided configuration).
+
 - Fix issue handling connection error during secure startup.
 
 1.8.0 (2023-12-14)